requirements.txt
asgiref==3.8.1
Babel==2.15.0
bleach==6.1.0
Django==5.0.6
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
django-summernote==0.8.20.0
mysqlclient==2.2.4
phonenumbers==8.13.37
pillow==10.3.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
webencodings==0.5.1